refactor ApplySIGImages to test pkg
