+++ pkg/test/nodeclass.go
@@ -19,13 +19,16 @@ package test
 import (
 	"context"
 	"fmt"
+	"sort"
 
 	"github.com/Azure/karpenter-provider-azure/pkg/apis/v1alpha2"
+	"github.com/Azure/karpenter-provider-azure/pkg/providers/imagefamily"
 	opstatus "github.com/awslabs/operatorpkg/status"
 	"github.com/blang/semver/v4"
 	"github.com/imdario/mergo"
 	"github.com/samber/lo"
 	corev1 "k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	karpv1 "sigs.k8s.io/karpenter/pkg/apis/v1"
@@ -123,6 +126,49 @@ func ApplyCIGImages(nodeClass *v1alpha2.AKSNodeClass) {
 			},
 		},
 	}
+}
+
+func ApplySIGImages(nodeClass *v1alpha2.AKSNodeClass) {
+	sigImageVersion := "202410.09.0"
+	imageFamilyNodeImages := getExpectedTestSIGImages(*nodeClass.Spec.ImageFamily, sigImageVersion)
+	nodeClass.Status.Images = translateToStatusNodeImages(imageFamilyNodeImages)
+}
+
+func getExpectedTestSIGImages(imageFamily string, version string) []imagefamily.NodeImage {
+	var images []imagefamily.DefaultImageOutput
+	if imageFamily == v1alpha2.Ubuntu2204ImageFamily {
+		images = imagefamily.Ubuntu2204{}.DefaultImages()
+	} else if imageFamily == v1alpha2.AzureLinuxImageFamily {
+		images = imagefamily.AzureLinux{}.DefaultImages()
+	}
+	nodeImages := []imagefamily.NodeImage{}
+	for _, image := range images {
+		nodeImages = append(nodeImages, imagefamily.NodeImage{
+			ID:           fmt.Sprintf("/subscriptions/10945678-1234-1234-1234-123456789012/resourceGroups/%s/providers/Microsoft.Compute/galleries/%s/images/%s/versions/%s", image.GalleryResourceGroup, image.GalleryName, image.ImageDefinition, version),
+			Requirements: image.Requirements,
+		})
+	}
+	return nodeImages
+}
+
+func translateToStatusNodeImages(imageFamilyNodeImages []imagefamily.NodeImage) []v1alpha2.NodeImage {
+	return lo.Map(imageFamilyNodeImages, func(nodeImage imagefamily.NodeImage, _ int) v1alpha2.NodeImage {
+		reqs := lo.Map(nodeImage.Requirements.NodeSelectorRequirements(), func(item karpv1.NodeSelectorRequirementWithMinValues, _ int) v1.NodeSelectorRequirement {
+			return item.NodeSelectorRequirement
+		})
+
+		// sorted for consistency
+		sort.Slice(reqs, func(i, j int) bool {
+			if len(reqs[i].Key) != len(reqs[j].Key) {
+				return len(reqs[i].Key) < len(reqs[j].Key)
+			}
+			return reqs[i].Key < reqs[j].Key
+		})
+		return v1alpha2.NodeImage{
+			ID:           nodeImage.ID,
+			Requirements: reqs,
+		}
+	})
 }
 
 func AKSNodeClassFieldIndexer(ctx context.Context) func(cache.Cache) error {
